services/camera-stream-grpc-app/requirements.txt
grpcio>=1.50.0
grpcio-health-checking>=1.50.0
grpcio-tools>=1.50.0
opencv-python>=4.7.0
pillow>=9.5.0
numpy>=1.24.0
protobuf>=4.21.0